u32 pointer to u8
OK

+++ include/data.h
@@ -87,7 +87,7 @@ struct Trainer
 #define TRAINER_ENCOUNTER_MUSIC(trainer)((gTrainers[trainer].encounterMusic_gender & 0x7F))
 
 extern const u16 gMinigameDigits_Pal[];
-extern const u32 gMinigameDigits_Gfx[];
+extern const u8 gMinigameDigits_Gfx[];
 
 extern const struct SpriteFrameImage gBattlerPicTable_PlayerLeft[];
 extern const struct SpriteFrameImage gBattlerPicTable_OpponentLeft[];
